backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.3
Flask-JWT-Extended==4.4.4
passlib[bcrypt]==1.7.4
requests==2.31.0
psycopg[binary]==3.1.15
openai>=1.0.0